Extract NativeLibrary polyfills to another class
